refactor(location): refactor entrance Locations
